Extract common types into own folder
