+++ packages/contract/src/common/responses.ts
@@ -1,0 +1,3 @@
+import z from 'zod';
+
+export const CommonResponseSchema = z.object({ id: z.string(), name: z.string() }); // Used for create, edit, delete responses
